requirements.txt
flask
azure-storage-queue
